quant-stack-india/requirements.txt
pandas==2.1.4
numpy==1.26.4
pandas-ta==0.3.14b0
yfinance==0.2.36
nsepython==2.9
jugaad-data==0.27
nsetools==1.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11
vectorbt==0.26.1
backtrader==1.9.78.123
pyfolio-reloaded==0.9.5
statsmodels==0.14.1
scikit-learn==1.4.0
torch==2.2.0
shap==0.44.0
optuna==3.5.0
PyPortfolioOpt==1.5.5
pykalman==0.9.7
kiteconnect==5.0.1
smartapi-python==1.3.9
fyers-apiv3==3.1.3
pyotp==2.9.0
selenium==4.18.1
APScheduler==3.10.4
python-telegram-bot==21.0.1
streamlit==1.31.0
plotly==5.18.0
fredapi==0.5.1
pytz==2024.1
scipy==1.12.0
openpyxl==3.1.2
SQLAlchemy==2.0.27
python-dotenv==1.0.1
numba==0.58.1
pytest==8.0.0
pytest-cov==4.1.0
holidays==0.45
colorama==0.4.6
pyyaml==6.0.1
